xtweetanalyser/requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.7
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
tweepy==4.14.0
tzdata==2024.2
urllib3==2.2.3
python-dotenv